Adding support for ZSH
The only incompatibility being the way ZSH handles enumerating strings.

To keep backward-compatibility, instead of changing the type of $BUNDLED_COMMANDS
to an array, we wrap the alias-setting part in a temporary function, and set the
`shwordsplit` ZSH option if inside a ZSH shell; then we get rid of that function
to avoid polluting the global environment.

Tested to work with both bash and zsh.